# openapi.yaml
openapi: 3.0.3
info:
  title: Movie Reservation System
  version: 1.0.0
  description: |
    Every successful response has the shape `{ message, data }`.
    Every error response has the shape `{ error, message? }` — `error` is a stable error code, `message` only appears when extra explanation is needed.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
...
    Theater:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Cinema 1 }
    Seat:
      type: object
      properties:
        id: { type: string, format: uuid }
        theaterId: { type: string, format: uuid }
        row: { type: string, example: A }
        number: { type: integer, example: 7 }
        type: { type: string, enum: [REGULAR, VIP] }
    Showtime:
      type: object
      properties:
        id: { type: string, format: uuid }
        movieId: { type: string, format: uuid }
        theaterId: { type: string, format: uuid }
        startsAt: { type: string, format: date-time }
        price:
          type: string
          description: Decimal(10,2), serialized as a string
          example: '95000'
        movie:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
        theater:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
...
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        totalPages: { type: integer, example: 3 }
    Error:
      type: object
      properties:
        error: { type: string, description: Stable error code, example: GenreNotFound }
        message: { type: string, description: Only present when extra explanation is needed }
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            issues:
              type: array
              description: Zod issue list, only present when error is ValidationError
              items: { type: object }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
    Unauthorized:
      description: Missing cookie, invalid token, or a token revoked by logout
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Unauthorized }
    Forbidden:
      description: Authenticated but not an ADMIN
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: Forbidden }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - cookieAuth: []
paths:
...
  /api/theaters:
    get:
      tags: [Theaters]
      summary: List theaters
      security: []
      parameters:
        - { name: name, in: query, description: 'Case-insensitive match on name', schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
        - { name: orderBy, in: query, description: Ordered by name, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Theater' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/theaters/{theaterId}:
    parameters:
      - { name: theaterId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Theaters]
      summary: Get a theater
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Theater' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/showtimes:
    get:
      tags: [Showtimes]
      summary: List showtimes
      security: []
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: movieId, in: query, schema: { type: string, format: uuid } }
        - { name: theaterId, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, description: 'startsAt >= from; accepts YYYY-MM-DD or an ISO date-time', schema: { type: string }, example: '2026-10-01' }
        - { name: to, in: query, description: 'startsAt <= to; accepts YYYY-MM-DD or an ISO date-time', schema: { type: string }, example: '2026-10-31' }
        - { name: orderBy, in: query, description: Ordered by startsAt, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Pagination'
                      - type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Showtime' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/showtimes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Showtimes]
      summary: Get a showtime
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/showtimes/{id}/seats:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Showtimes]
      summary: The seat map of a showtime's theater
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      movie:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          title: { type: string }
                      theater:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          name: { type: string }
                      showtime: { type: string, format: date-time, description: The showtime's startsAt }
                      seats:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Seat'
                            - type: object
                              properties:
                                reserved:
                                  type: boolean
                                  description: Already booked for THIS showtime — the same seat is free at other showtimes
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/theaters:
    post:
      tags: [Admin]
      summary: Create a theater
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Cinema 1 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Theater' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: TheaterAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: TheaterAlreadyExists }
  /api/admin/theaters/{theaterId}:
    parameters:
      - { name: theaterId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Rename a theater
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: Cinema 2 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Theater' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: TheaterAlreadyExists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: TheaterAlreadyExists }
...
  /api/admin/showtimes:
    post:
      tags: [Admin]
      summary: Create a showtime
      description: A theater cannot have two showtimes with the same start time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [movieId, theaterId, startsAt, price]
              properties:
                movieId: { type: string, format: uuid }
                theaterId: { type: string, format: uuid }
                startsAt: { type: string, format: date-time, example: '2026-10-01T19:00:00Z' }
                price: { type: number, minimum: 0, exclusiveMinimum: true, example: 95000 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '400':
          description: '`MovieNotFound` or `TheaterNotFound` (missing or soft-deleted)'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: MovieNotFound }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: ShowtimeConflict — the theater already has a showtime at this exact start time
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: ShowtimeConflict }
  /api/admin/showtimes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Admin]
      summary: Update a showtime — at least one field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movieId: { type: string, format: uuid }
                theaterId: { type: string, format: uuid }
                startsAt: { type: string, format: date-time }
                price: { type: number, example: 120000 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '400':
          description: '`EmptyUpdate`, `MovieNotFound` or `TheaterNotFound`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: EmptyUpdate, message: At least one field is required }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: ShowtimeConflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Admin]
      summary: Delete a showtime
      description: A hard delete, not a soft delete. Refused once the showtime has reservations.
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Showtime' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: ShowtimeHasReservations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: ShowtimeHasReservations, message: 2 reservation(s) already booked }
...